Ooops, the tests broke after the last refactor.  Fix typo in amountFor return type.

+++ aFirstExample/src/main/groovy/Customer.groovy
@@ -53,7 +53,7 @@
 		result
 	}
 
-	private int amountFor(Rental aRental){
+	private double amountFor(Rental aRental){
 
 		double thisAmount = 0
 		//determine amounts for each line
